Extract AP code to module
