+++ src/board/wifi/mod.rs
@@ -1,25 +1,21 @@
 use crate::{
-    board::hal::{
-        clock::Clocks,
-        peripherals::{RNG, TIMG1},
-        radio::Wifi,
-        system::{PeripheralClockControl, RadioClockControl},
-        timer::{Timer0, TimerGroup},
-        Rng, Timer,
+    board::{
+        hal::{
+            clock::Clocks,
+            peripherals::{RNG, TIMG1},
+            radio::Wifi,
+            system::{PeripheralClockControl, RadioClockControl},
+            timer::{Timer0, TimerGroup},
+            Rng, Timer,
+        },
+        wifi::ap::ApState,
     },
     replace_with::replace_with_or_abort_async,
     task_control::TaskController,
 };
-use embassy_executor::Spawner;
-use embassy_futures::join::join;
 use embassy_net::{Config, Stack, StackResources};
-use embassy_sync::{blocking_mutex::raw::NoopRawMutex, mutex::Mutex};
 use embedded_hal_old::prelude::_embedded_hal_blocking_rng_Read;
-use embedded_svc::wifi::{AccessPointConfiguration, Configuration, Wifi as _};
-use esp_wifi::{
-    wifi::{WifiController, WifiDevice, WifiEvent, WifiMode, WifiState},
-    EspWifiInitFor, EspWifiInitialization,
-};
+use esp_wifi::{wifi::WifiDevice, EspWifiInitFor, EspWifiInitialization};
 use rand_core::{RngCore, SeedableRng};
 use replace_with::replace_with_or_abort;
 use wyhash::WyRng;
@@ -32,82 +28,12 @@ pub unsafe fn as_static_mut<T>(what: &mut T) -> &'static mut T {
     core::mem::transmute(what)
 }
 
+pub mod ap;
+
 pub struct WifiDriver {
     wifi: Wifi,
     rng: WyRng,
     state: WifiDriverState,
-}
-
-struct ApState {
-    _init: EspWifiInitialization,
-    controller: WifiController<'static>,
-    stack: Stack<WifiDevice<'static>>,
-    connection_task_control: TaskController<()>,
-    net_task_control: TaskController<()>,
-    client_count: Mutex<NoopRawMutex, u32>,
-    started: bool,
-}
-
-impl ApState {
-    fn new(
-        _init: EspWifiInitialization,
-        config: Config,
-        wifi: &'static mut Wifi,
-        resources: &'static mut StackResources<3>,
-        random_seed: u64,
-    ) -> Self {
-        let (wifi_interface, controller) =
-            esp_wifi::wifi::new_with_mode(&_init, wifi, WifiMode::Ap);
-
-        Self {
-            _init,
-            controller,
-            stack: Stack::new(wifi_interface, config, resources, random_seed),
-            connection_task_control: TaskController::new(),
-            net_task_control: TaskController::new(),
-            client_count: Mutex::new(0),
-            started: false,
-        }
-    }
-
-    async fn start(&mut self) -> &mut Stack<WifiDevice<'static>> {
-        if !self.started {
-            let spawner = Spawner::for_current_executor().await;
-            unsafe {
-                spawner.must_spawn(ap_task(
-                    as_static_mut(&mut self.controller),
-                    as_static_ref(&self.connection_task_control),
-                    as_static_ref(&self.client_count),
-                ));
-                spawner.must_spawn(net_task(
-                    as_static_ref(&self.stack),
-                    as_static_ref(&self.net_task_control),
-                ));
-            }
-            self.started = true;
-        }
-
-        &mut self.stack
-    }
-
-    async fn stop(&mut self) {
-        if self.started {
-            let _ = join(
-                self.connection_task_control.stop_from_outside(),
-                self.net_task_control.stop_from_outside(),
-            )
-            .await;
-            self.started = false;
-        }
-    }
-
-    fn is_running(&self) -> bool {
-        !self.connection_task_control.has_exited() && !self.net_task_control.has_exited()
-    }
-
-    async fn client_count(&self) -> u32 {
-        *self.client_count.lock().await
-    }
 }
 
 #[allow(clippy::large_enum_variant)]
@@ -222,72 +148,6 @@ pub async fn net_task(
     task_control
         .run_cancellable(async {
             stack.run().await;
-        })
-        .await;
-}
-
-#[embassy_executor::task]
-pub async fn ap_task(
-    controller: &'static mut WifiController<'static>,
-    task_control: &'static TaskController<()>,
-    client_count: &'static Mutex<NoopRawMutex, u32>,
-) {
-    task_control
-        .run_cancellable(async {
-            log::info!("Start connection task");
-            log::debug!("Device capabilities: {:?}", controller.get_capabilities());
-
-            loop {
-                if !matches!(controller.is_started(), Ok(true)) {
-                    let client_config = Configuration::AccessPoint(AccessPointConfiguration {
-                        ssid: "Card/IO".into(),
-                        max_connections: 1,
-                        ..Default::default()
-                    });
-                    controller.set_configuration(&client_config).unwrap();
-                    log::info!("Starting wifi");
-
-                    controller.start().await.unwrap();
-                    log::info!("Wifi started!");
-                }
-
-                if let WifiState::ApStart
-                | WifiState::ApStaConnected
-                | WifiState::ApStaDisconnected = esp_wifi::wifi::get_wifi_state()
-                {
-                    let events = controller
-                        .wait_for_events(
-                            WifiEvent::ApStop
-                                | WifiEvent::ApStaconnected
-                                | WifiEvent::ApStadisconnected,
-                            false,
-                        )
-                        .await;
-
-                    if events.contains(WifiEvent::ApStaconnected) {
-                        let count = {
-                            let mut count = client_count.lock().await;
-                            *count = count.saturating_add(1);
-                            *count
-                        };
-                        log::info!("Client connected, {count} total");
-                    }
-                    if events.contains(WifiEvent::ApStadisconnected) {
-                        let count = {
-                            let mut count = client_count.lock().await;
-                            *count = count.saturating_sub(1);
-                            *count
-                        };
-                        log::info!("Client disconnected, {count} left");
-                    }
-                    if events.contains(WifiEvent::ApStop) {
-                        log::info!("AP stopped");
-                        return;
-                    }
-
-                    log::info!("Event processing done");
-                }
-            }
         })
         .await;
 }
